Change tasks display to use Resource Cards
