+++ ui/src/tasks/containers/TasksPage.tsx
@@ -10,6 +10,7 @@ import {Page} from 'src/pageLayout'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 import ImportOverlay from 'src/shared/components/ImportOverlay'
 import FilterList from 'src/shared/components/Filter'
+import SearchWidget from 'src/shared/components/search_widget/SearchWidget'
 
 // Actions
 import {
@@ -101,11 +102,10 @@ class TasksPage extends PureComponent<Props, State> {
         <Page titleTag="Tasks">
           <TasksHeader
             onCreateTask={this.handleCreateTask}
-            setSearchTerm={setSearchTerm}
             setShowInactive={setShowInactive}
             showInactive={showInactive}
             toggleOverlay={this.handleToggleImportOverlay}
-            searchTerm={searchTerm}
+            filterComponent={() => this.search}
           />
           <Page.Contents fullWidth={false} scrollable={true}>
             <div className="col-xs-12">
@@ -128,6 +128,7 @@ class TasksPage extends PureComponent<Props, State> {
                     onRemoveTaskLabels={onRemoveTaskLabels}
                     onRunTask={onRunTask}
                     onFilterChange={setSearchTerm}
+                    filterComponent={() => this.search}
                   />
                 )}
               </FilterList>
@@ -165,6 +166,18 @@ class TasksPage extends PureComponent<Props, State> {
 
   private handleToggleImportOverlay = (): void => {
     this.setState({isImporting: !this.state.isImporting})
+  }
+
+  private get search(): JSX.Element {
+    const {setSearchTerm, searchTerm} = this.props
+
+    return (
+      <SearchWidget
+        placeholderText="Filter tasks..."
+        onSearch={setSearchTerm}
+        searchTerm={searchTerm}
+      />
+    )
   }
 
   private get importOverlay(): JSX.Element {
